Split actions index into multiple files.
